% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_spline_basis.R
\name{bs_bbase_precise}
\alias{bs_bbase_precise}
\title{Small wrapper to create a B-spline basis similar to the original bs_bbase_precise}
\usage{
bs_bbase_precise(
  x = x,
  lastobs = max(x),
  xl = min(x),
  xr = max(x),
  nseg = 10,
  deg = 3
)
}
\arguments{
\item{x}{numeric vector of times to evaluate the basis}

\item{lastobs}{numeric last observed year (upper knot placement)}

\item{nseg}{integer number of interior segments (default 10)}
}
\value{
list(B.ik = basis matrix (length(all_years) x K), Kstar = effective K, knots.k = knots)
}
\description{
Small wrapper to create a B-spline basis similar to the original bs_bbase_precise
}
\keyword{internal}
