% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country_codes_DHS.R
\docType{data}
\name{country_codes_DHS}
\alias{country_codes_DHS}
\title{Country Codes for DHS Surveys}
\format{
A data frame with 91 rows and 4 variables:
\describe{
\item{Country Name}{DHS country name}
\item{Code}{2-letter ISO country code}
\item{India States}{State codes of India}
\item{State Name}{State names of India}
}
}
\source{
"DHS Program"
}
\usage{
data(country_codes_DHS)
}
\description{
A dataset with ISO and DHS-specific country codes used in the
\code{mcmsector} package.
}
\keyword{datasets}
