% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_mcmsector_params.R
\name{extract_mcmsector_params}
\alias{extract_mcmsector_params}
\title{Extract alpha_pms and beta.k MCMC samples from a JAGS model object}
\usage{
extract_mcmsector_params(mod, n_method, n_subnat, n_beta = 13)
}
\arguments{
\item{mod}{A JAGS model object loaded with \code{readRDS()}, containing
\code{mod$BUGSoutput$sims.array}.}

\item{n_method}{Integer; number of contraceptive methods.}

\item{n_subnat}{Integer; number of subnational regions.}

\item{n_beta}{Integer; number of beta coefficients (typically 13).}
}
\value{
A list with:
\describe{
\item{alpha_pms}{A matrix of posterior samples for alpha parameters.}
\item{beta_k}{A matrix of posterior samples for beta parameters.}
}
}
\description{
Given a JAGS model object containing \code{BUGSoutput$sims.array}, this function
extracts the posterior simulations for \code{alpha_pms} and \code{beta.k} into
clean matrices suitable for downstream computation.
}
\examples{
\dontrun{
mod <- readRDS("results/JAGS/JAGS_model.RDS")
params <- extract_mcmsector_params(mod, n_method=5, n_subnat=159, n_beta=13)
}
}
