% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_region_names.R
\name{fix_region_names}
\alias{fix_region_names}
\title{Harmonise Region Names for Selected Countries}
\usage{
fix_region_names(df)
}
\arguments{
\item{df}{A data frame containing \code{Country} and \code{Region} columns.}
}
\value{
A data frame with modified region names.
}
\description{
This function standardises region names for Burkina Faso, Rwanda,
Nigeria, and Cote d'Ivoire to ensure consistent naming across years.
}
\examples{
clean <- fix_region_names(mcmsector::subnat_bivar_data)

}
