% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subnat_bivar_data.R
\docType{data}
\name{subnat_bivar_data}
\alias{subnat_bivar_data}
\title{DHS survey observations for the proportion of modern contraceptives supplied by the public and private sectors at the subnational level}
\format{
\subsection{\code{subnat_bivar_data}}{

A data frame with 1737 rows and 12 columns:
\describe{
\item{Country}{Country names}
\item{Region}{Regional names}
\item{Method}{Contraceptive method name}
\item{average_year}{Average year of the survey}
\item{year}{Year of the survey}
\item{country_code}{ISO country codes}
\item{Public}{Proportion supplied by the Public sector}
\item{se.Public}{Standard error of proportion supplied by the Public sector}
\item{Public_n}{Sample size used to calculate proportion supplied by the Public sector}
\item{Private}{Proportion supplied by the private sector}
\item{se.Private}{Standard error of proportion supplied by the private sector}
\item{Private_n}{Sample size used to calculate proportion supplied by the private sector}
\item{check_sum}{Check to see if proportions sum to 1}
}
}
}
\source{
On request from DHS microdata - using the womens individuals recode file. Contact details found at https://dhsprogram.com/data/dataset_admin/login_main.cfm
}
\usage{
subnat_bivar_data
}
\description{
DHS survey observations for the proportion of modern contraceptives supplied by the public and private sectors at the subnational level
}
\keyword{datasets}
