% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_posterior_proportions.R
\name{simulate_posterior_proportions}
\alias{simulate_posterior_proportions}
\title{Simulate posterior public/private proportions from alpha and beta parameters}
\usage{
simulate_posterior_proportions(
  alpha_pms,
  beta_k,
  B_ik,
  n_method,
  n_subnat,
  all_years,
  n_samps = 4000
)
}
\arguments{
\item{alpha_pms}{Matrix of alpha posterior samples.}

\item{beta_k}{Matrix of beta posterior samples.}

\item{B_ik}{A 3D array of basis functions with dimensions:
(subnational, year, n_beta).}

\item{n_method}{Number of methods.}

\item{n_subnat}{Number of subnational units.}

\item{all_years}{Vector of years included in the model.}

\item{n_samps}{Number of posterior samples to compute (default = 4000).}
}
\value{
A list containing:
\describe{
\item{P}{A 5D array of public/private probabilities.}
}
}
\description{
Computes posterior draws of the logit-scale proportions and the corresponding
public/private probabilities for all methods, regions, and years.
}
